+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `Mode` protocol
+
 ### Breaking
 
 - Rename `Chord` to `ChordVoicing`
+- Rename `Mode` to `MajorMode`
 
 ## [1.0.0] - 2020-07-11